Got the lists, on the C side working
On the cSide, the types are now RRList (heterogenous list) and RRStringArray (array of strings). No need for the "c" in cRRList. A client will use this from 'C' so there can not be anything else than a "C" list.

The core of the C-api is C++, so malloc should not be used, only new/delete. The use of realloc in the list structures need to be modified.